rework toml-reader as print_components within /ci folder
