+++ .gitignore
@@ -2,6 +2,7 @@
 # will have compiled files and executables
 /target/
 /component/target/
+/ci/print_components/target/
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
