+++ ci/print_components/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "toml-reader"
+name = "print_components"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,6 +7,4 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-toml_edit = { version = "0.13", features = ["serde", "easy"] }
+component = { path = "../../component" }
